Implement the status-update-service
This interface has been updated in a rather complicated way because we are using a chain itself being an implementation of the same interface. It was just thought complicated as it is it would make the size of the implementation objects smaller and easy to test, at the same time it would make the implementation structure robust and also easy to add features in future. For instance it's thought data uploaded by file will continue to be added to the project, therefore making it necessary to have more batch structured as compilation of formerly compressed datasets and that will mean continual enhancement of the entire architecture. It is designed to handle such improvements with less fuss, therefore even the list of services itself is configured independently in a spring container. The spirit of this over-engineering is to make future development easier